move check pypi release to utils
